Initial push for legacy-api
I dubbed this branch "legacy-api" because it attempts to implement
the required api endpoints that are required to run the legacy
CSB application pretty much unchanged, with only some minor changes
to make the endpoints follow something that makes more sense in this
context. Those are in api/api.php along with a .htaccess file that
aims the endpoints to the file.

If it gets expanded and needs to get split up, you can do that by
pointing some of the endpoints to a different file (obviously).

This changes quite a few files, but the mapper applications shouldn't
show any missing files now (hopefully!).

What I also changed was that the application definitions themselves
all reside in a separate self-contained folder under csb-apps. That
probably duplicates a few icons, but on the other hand you can freely
change stuff without worrying about other applications suddently not
finding their images any more, for example.

Changed as well is that the "app" parameter now uses the same name than
is stored in the database (a benefit that probably shouldn't need too
much of an explanation), even though the app menu and nav items aren't
generated automagically yet.

There is still a whole lot of applications and/or images in the folder
in csb-content, but I'll leave that for a future push.

LMK if I didn't think anything through completely.